first init for ure-assist.php
This script assumes the User Role "webmaster" and removes the Additional Capabilities drop down from the users "admin" and "administrator".  The script also loads a small css file that hides the "capabilities" link in the user.php page. This css file is is needed since ure_manage_options has to be turned on in order for the User Role to be able to edit Users, and in doing so, it adds the "capabilities" link next to each user in the users.php page.